Make CompositeGlyphs* usable.
render.xml no longer describes the CompositeGlyphs* requests as taking
lists of complicated unions of structures of lists: it says instead that
they take a LISTofBYTE. The caller is responsible for constructing an
appropriate sequence of glyph elements. Previously, the requests could
not actually be used because XCB did not correctly compute the length of
the provided data.