Corrected type of FakeInput rootX & rootY fields
The FakeInput request defined by the XTEST extension has two fields, rootX and
rootY, that are used to indicate relative or absolute pointer coordinates for
fake motion events. These fields are of type INT16, not CARD16. (They must be
signed to allow for negative relative motions and to align with the root-x and
root-y fields of MotionNotify events.)

Signed-off-by: Peter Harris <pharris@opentext.com>