xinput: enum InputClass: comment value Attach 7
The value Attach=7 is defined in XI.h but not supported anymore by libXi.
And it never was transmitted over the protocol.
(It was used to tag a fake class created by libXi)

Add this comment so that future reviews will not accidentally add this value.

spec:
http://cgit.freedesktop.org/xorg/proto/inputproto/tree/XI.h#n242

drop of support in libXi:
http://cgit.freedesktop.org/xorg/lib/libXi/commit/?id=62858c426e6de1b99df660251737233afd335302