present: specify length of RedirectNotify.notifies
The list "notifies" of the event RedirectNotify didn't have
a length.

The missing length has caused uncompilable C-Code with our
event-acessors patch because the length-acessor for that list
has used an undeclared variable for the list-length.

Having a length for that list is good anyways for ease of use.

Since the event doesn't contain a field which specifies the
length of the list, the length is derived from the length
of the event.

spec:
http://cgit.freedesktop.org/xorg/proto/presentproto/tree/presentproto.txt?id=presentproto-1.0#n429
http://cgit.freedesktop.org/xorg/proto/presentproto/tree/presentproto.txt?id=presentproto-1.0#n710

note:
The event definition starting at line 429 lists the relationship between
length of event ( "length" ) and length of list ( "n" )
as 17+2n where as the definition starting at line 710
lists 18+2n. The latter is correct according to my verification.
The spec should be fixed.

Message-ID: <1409845742-38797-1-git-send-email-chris@demorecorder.com>
Patch-Thread-Subject: [Xcb] support popcount of a list and associated xml changes
Patch-Set: PopcountList
Patch-Number: proto 1/8
Patch-Version: V1
Signed-off-by: Christian Linhart <chris@DemoRecorder.com>