Replace valueparam with CARD32 bit field in Xprint
Replaced valueparam with a simple 32 bit field for PrintSelectInput
and PrintInputSelected by looking at the
lib code at :
http://cgit.freedesktop.org/xorg/lib/libXp/tree/src/XpInput.c?id=libXp-1.0.2

Signed-off-by: Jaya Tiwari <tiwari.jaya18@gmail.com>
Reviewed-by: Christian Linhart <chris@demorecorder.com>

Review-comments by Christian Linhart:
This patch fixes a bug:
The prior use of valueparam was incorrect, and you have fixed that bug with this patch.

This can be verified by looking at the code which you have linked to.
and by looking at the definitions of the protocol structs:
    For the PrintSelectInput-request:
        http://cgit.freedesktop.org/xorg/proto/printproto/tree/Printstr.h?id=printproto-1.0.5#n439
    For the reply of PrintInputSelected:
        http://cgit.freedesktop.org/xorg/proto/printproto/tree/Printstr.h?id=printproto-1.0.5#n457