fix ABI/API compatibility problem in present
for function xcb_present_redirect_notify_sizeof

removing the explicit length of the list that was added since the
last release with the following changeset:
http://cgit.freedesktop.org/xcb/proto/commit/?id=c9b1523b23af52087a6354730f86b8d19fa3c7c0

Removing the explicit length again is OK because
implicit length is allowed when a list is the last
field of a request or event.

The compile problem that the prior change has tried to fix,
has been fixed in another way. Therefore removing the explicit length
is safe now.

Signed-off-by: Christian Linhart <chris@demorecorder.com>