xinput: typedef for event_type_base
member event_type_base in struct InputClassInfo is a CARD8.
This patch replaces the type with the new typedef EventTypeBase
which is defined as a CARD8.

This does not alter the ABI or API for that struct because
typedefs do not appear in the ABI.
And C handles typedefs as aliases.
So there'll be no API issues as well.

The libxcb generator adds additional functions that implement
an iterator over this type.
This is harmless because adding code is ABI/API compatible.

The purpose of this change is to add more semantic info to this
member variable of this struct.

This helps with xcb-based generators or automatic code-checkers
and things like that.

Signed-off-by: Christian Linhart <chris@demorecorder.com>