Release xcb-proto 1.13
This release adds support for variable-sized lists of FDs (e.g. sized
by a field in the request/event), as well as several other cleanups and
fixes in the core code.

DRI3 has been bumped to v1.2, using the new support to support
multi-planar buffers and DRM modifiers.

Present has also been bumped to v1.2, adding a new token for the server
to indicate to the client that it should reallocate its buffers.

RandR has been updated to 1.6, adding support for DRM leases to allow
clients to directly control outputs.

Signed-off-by: Daniel Stone <daniels@collabora.com>